remove tendermint dependency (again)
