+++ types/proofs/row_proof.go
@@ -1,0 +1,15 @@
+package proofs
+
+import "github.com/celestiaorg/go-square/merkle"
+
+// RowProof is a Merkle proof that a set of rows exist in a Merkle tree with a
+// given data root.
+type RowProof struct {
+	// RowRoots are the roots of the rows being proven.
+	RowRoots []byte `json:"row_roots"`
+	// Proofs is a list of Merkle proofs where each proof proves that a row
+	// exists in a Merkle tree with a given data root.
+	Proofs   []*merkle.Proof `json:"proofs"`
+	StartRow uint32          `json:"start_row"`
+	EndRow   uint32          `json:"end_row"`
+}
